Moved mailer processor in common processors. Renamed it SendMail.
