+++ logisland-assembly/pom.xml
@@ -95,10 +95,6 @@
             <artifactId>logisland-bro-plugin</artifactId>
         </dependency>
         <dependency>
-            <groupId>com.hurence.logisland</groupId>
-            <artifactId>logisland-mailer-plugin</artifactId>
-        </dependency>
-        <dependency>
             <groupId>org.apache.commons</groupId>
             <artifactId>commons-email</artifactId>
         </dependency>
