+++ logisland-framework/logisland-documentation/pom.xml
@@ -77,10 +77,6 @@
             <groupId>com.hurence.logisland</groupId>
             <artifactId>logisland-scripting-processors-plugin</artifactId>
         </dependency>
-        <dependency>
-            <groupId>com.hurence.logisland</groupId>
-            <artifactId>logisland-mailer-plugin</artifactId>
-        </dependency>
     </dependencies>
     <build>
         <plugins>
