--- logisland-plugins/logisland-mailer-plugin/src/test/resources/log4j.properties
@@ -1,36 +1,0 @@
-#
-# Licensed to the Apache Software Foundation (ASF) under one or more
-# contributor license agreements.  See the NOTICE file distributed with
-# this work for additional information regarding copyright ownership.
-# The ASF licenses this file to You under the Apache License, Version 2.0
-# (the "License"); you may not use this file except in compliance with
-# the License.  You may obtain a copy of the License at
-#
-#    http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-# Set everything to be logged to the console
-log4j.rootCategory=WARN, console
-log4j.appender.console=org.apache.log4j.ConsoleAppender
-log4j.appender.console.target=System.err
-log4j.appender.console.layout=org.apache.log4j.PatternLayout
-log4j.appender.console.layout.ConversionPattern=%d{yy/MM/dd HH:mm:ss} %p %c{1}: %m%n
-
-#org.apache.zookeeper.server.ZooKeeperServer
-log4j.logger.org.apache.spark=WARN
-log4j.logger.org.apache.spark.scheduler=WARN
-log4j.logger.org.apache.spark.history=WARN
-log4j.logger.org.spark-project.jetty=WARN
-log4j.logger.io.netty=WARN
-log4j.logger.org.apache.zookeeper=WARN
-log4j.logger.org.apache.hadoop.ipc.Client=WARN
-log4j.logger.org.apache.hadoop=WARN
-log4j.logger.org.apache.hadoop.ipc.ProtobufRpcEngine=WARN
-log4j.logger.parquet.hadoop=WARN
-log4j.logger.com.hurence=DEBUG
