+++ pom.xml
@@ -601,11 +601,6 @@
             </dependency>
             <dependency>
                 <groupId>com.hurence.logisland</groupId>
-                <artifactId>logisland-mailer-plugin</artifactId>
-                <version>${project.version}</version>
-            </dependency>
-            <dependency>
-                <groupId>com.hurence.logisland</groupId>
                 <artifactId>logisland-spark-engine21</artifactId>
                 <version>${project.version}</version>
             </dependency>
@@ -642,11 +637,6 @@
             <dependency>
                 <groupId>com.hurence.logisland</groupId>
                 <artifactId>logisland-agent</artifactId>
-                <version>${project.version}</version>
-            </dependency>
-            <dependency>
-                <groupId>com.hurence.logisland</groupId>
-                <artifactId>logisland-mailer-plugin</artifactId>
                 <version>${project.version}</version>
             </dependency>
             <dependency>
